Added input method key switching (alt+space) Added code to enable deskbar replicant loading after input_server is started
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@10664 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/servers/input/InputServer.h
@@ -166,7 +166,7 @@ public:
 	static status_t StartStopDevices(BInputServerDevice *isd, bool);
 	static status_t ControlDevices(const char *, input_device_type, unsigned long, BMessage*);
 
-	bool DoMouseAcceleration(long*, long*);
+	bool DoMouseAcceleration(int32*, int32*);
 	bool SetMousePos(long*, long*, long, long);
 	bool SetMousePos(long*, long*, BPoint);
 	bool SetMousePos(long*, long*, float, float);
